ventilairsec-vmi/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
python-enocean==0.61.3
paho-mqtt==1.6.1
